+++ velox/functions/sparksql/DecimalArithmetic.cpp
@@ -672,29 +672,6 @@ void registerDecimalDivide(
       LongDecimal<P1, S1>,
       ShortDecimal<P2, S2>>({functionName}, constraints);
 }
-
-template <typename TExec>
-struct DecimalCeilFunction {
-  VELOX_DEFINE_FUNCTION_TYPES(TExec);
-
-  template <typename A>
-  void initialize(
-      const std::vector<TypePtr>& inputTypes,
-      const core::QueryConfig& /*config*/,
-      A* /*a*/) {
-    rescaleFactor_ = velox::DecimalUtil::kPowersOfTen
-        [getDecimalPrecisionScale(*inputTypes[0]).second];
-  }
-
-  template <typename R, typename A>
-  void call(R& out, const A& a) {
-    const auto increment = (a % rescaleFactor_) > 0 ? 1 : 0;
-    out = a / rescaleFactor_ + increment;
-  }
-
- private:
-  int128_t rescaleFactor_;
-};
 } // namespace
 
 void registerDecimalAdd(const std::string& prefix) {
@@ -737,34 +714,5 @@ void registerDecimalDivide(const std::string& prefix) {
   registerDecimalDivide<DivideFunctionDenyPrecisionLoss>(
       prefix + "divide" + kDenyPrecisionLoss,
       getDivideConstraintsDenyPrecisionLoss());
-}
-
-void registerDecimalCeil(const std::string& prefix) {
-  std::vector<exec::SignatureVariable> constraints = {
-      exec::SignatureVariable(
-          P2::name(),
-          fmt::format(
-              "min(38, {p} - {s} + min({s}, 1))",
-              fmt::arg("p", P1::name()),
-              fmt::arg("s", S1::name())),
-          exec::ParameterType::kIntegerParameter),
-      exec::SignatureVariable(
-          S2::name(), "0", exec::ParameterType::kIntegerParameter),
-  };
-
-  registerFunction<
-      DecimalCeilFunction,
-      LongDecimal<P2, S2>,
-      LongDecimal<P1, S1>>({prefix + "ceil"}, constraints);
-
-  registerFunction<
-      DecimalCeilFunction,
-      ShortDecimal<P2, S2>,
-      LongDecimal<P1, S1>>({prefix + "ceil"}, constraints);
-
-  registerFunction<
-      DecimalCeilFunction,
-      ShortDecimal<P2, S2>,
-      ShortDecimal<P1, S1>>({prefix + "ceil"}, constraints);
 }
 } // namespace facebook::velox::functions::sparksql
